+++ changelog.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Module for getting and creating journal notes in Nova.
 - Tests for journal notes.
+- Module for accessing site and file endpoints in Microsoft Graph.
+- Minor refactoring to move common HTTP request wrappers for Microsoft Graph into their own file.
 
 ## [2.1.1] - 2024-04-10